pyproject.toml
[build-system]
requires = ["hatchling==1.25.0"]
build-backend = "hatchling.build"

[project]
name = "conciliador-bancario"
version = "0.1.0"
description = "CLI local para conciliacion bancaria (Chile-first) con auditoria y enfoque fail-closed."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Conciliador Bancario Contributors" }]
keywords = ["conciliacion", "bancaria", "chile", "contabilidad", "cli"]
dependencies = [
  "pydantic==2.7.4",
  "typer==0.12.3",
  "rich==13.7.1",
  "PyYAML==6.0.1",
  "openpyxl==3.1.2",
  "pypdf==4.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.2.0",
  "hypothesis==6.108.4",
  "ruff==0.4.10",
  "mypy==1.10.0",
  "types-PyYAML==6.0.12.20240808",
]
pdf_ocr = [
  # Requiere dependencias de sistema (poppler) para convertir PDF->imagenes.
  "pdf2image==1.17.0",
  "pytesseract==0.3.10",
  "Pillow==10.4.0",
]

[project.scripts]
concilia = "conciliador_bancario.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/conciliador_bancario", "src/conciliador_core"]

[tool.hatch.build]
include = [
  "src/conciliador_bancario/templates/*.yaml",
  "src/conciliador_bancario/templates/*.csv",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_equality = true
